Lab - 7 start

+++ src/main/java/lab/model/Country.java
@@ -1,13 +1,13 @@
 package lab.model;
 
 public interface Country {
-    int getId();
+    long getId();
 
     String getName();
 
     String getCodeName();
 
-    SimpleCountry setId(int id);
+    SimpleCountry setId(long id);
 
     SimpleCountry setName(String name);
